+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - [oclif] `oclif-plugin-spaced-commands` to fix alias documentation.
 
+## Fixed
+- [vtex test e2e] Screenshot and video assets updated according to tester-hub.
+
 ## [2.99.3] - 2020-05-05
 ### Fixed
 - [telemetry] Sanitize jwt tokens and simplify buffers serializtion.